Rename Page and UpgradablePackages
